refactor(auto-reply): split reply agent runner
